+++ eras/allegra/impl/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### `testlib`
 
 * Consolidate all `Arbitrary` instances from the test package to under a new `testlib`. #3285
+* Remove redundant pattern synonym `AllegraTxAuxData'`
 
 ## 1.0.0.0